+++ CHANGELOG.md
@@ -20,7 +20,7 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 - Support for Python 3.14 and Python 3.13.4.
   [#723](https://github.com/hynek/structlog/pull/723)
 
-- `structlog.tracebacks` now handles exception groups.
+- `structlog.tracebacks` now handles [exception groups](https://docs.python.org/3/library/exceptions.html#exception-groups).
   `structlog.tracebacks.Stack` has two new fields, `is_group: bool` and `exceptions: list[Trace]`.
   This works similarly to what Rich v14.0.0 does.
   [#720](https://github.com/hynek/structlog/pull/720)
@@ -37,7 +37,6 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 
 - `structlog.processors.TimeStamper` now again uses timestamps using UTC for custom format strings when `utc=True`.
   [#713](https://github.com/hynek/structlog/pull/713)
-
 
 
 ## [25.2.0](https://github.com/hynek/structlog/compare/25.1.0...25.2.0) - 2025-03-11